// backend/services/libs/shared/src/enum/activity.state.enum.ts
export enum ActivityStateEnum {
    MONITORING_REPORT_UPLOADED = 'MONITORING_REPORT_UPLOADED',
    MONITORING_REPORT_VERIFIED = 'MONITORING_REPORT_VERIFIED',
    MONITORING_REPORT_REJECTED = 'MONITORING_REPORT_REJECTED',
    VERIFICATION_REPORT_UPLOADED = 'VERIFICATION_REPORT_UPLOADED',
    VERIFICATION_REPORT_VERIFIED = 'VERIFICATION_REPORT_VERIFIED',
    VERIFICATION_REPORT_REJECTED = 'VERIFICATION_REPORT_REJECTED',
}