// backend/services/libs/shared/src/enum/project.audit.log.type.enum.ts
export enum ProjectAuditLogType {
  PENDING = "PENDING",
  REJECTED = "REJECTED",
  APPROVED = "APPROVED",
  PDD_SUBMITTED = "PDD_SUBMITTED",
  PDD_REJECTED_BY_CERTIFIER = "PDD_REJECTED_BY_CERTIFIER",
  PDD_APPROVED_BY_CERTIFIER = "PDD_APPROVED_BY_CERTIFIER",
  PDD_REJECTED_BY_DNA = "PDD_REJECTED_BY_DNA",
  PDD_APPROVED_BY_DNA = "PDD_APPROVED_BY_DNA",
  VALIDATION_REPORT_SUBMITTED = "VALIDATION_REPORT_SUBMITTED",
  VALIDATION_DNA_REJECTED = "VALIDATION_REPORT_REJECTED",
  VALIDATION_DNA_APPROVED = "VALIDATION_REPORT_APPROVED",
  AUTHORIZED = "AUTHORIZED",
  MONITORING_CREATE = "MONITORING_CREATE",
  MONITORING_APPROVED = "MONITORING_APPROVED",
  MONITORING_REJECTED = "MONITORING_REJECTED",
  VERIFICATION_CREATE = "VERIFICATION_CREATE",
  VERIFICATION_APPROVED = "VERIFICATION_APPROVED",
  VERIFICATION_REJECTED = "VERIFICATION_REJECTED",
  CREDIT_ISSUED = "CREDIT_ISSUED",
  TRANSFER_REQUESTED = "TRANSFER_REQUESTED",
  RETIRE_REQUESTED = "RETIRE_REQUESTED",
  TRANSFER_APPROVED = "TRANSFER_APPROVED",
  RETIRE_APPROVED = "RETIRE_APPROVED",
  TRANSFER_REJECTED = "TRANSFER_REJECTED",
  RETIRE_REJECTED = "RETIRE_REJECTED",
  TRANSFER_CANCELLED = "TRANSFER_CANCELLED",
  RETIRE_CANCELLED = "RETIRE_CANCELLED",
  MONITORING_REPORT_SUBMITTED = "MONITORING_REPORT_SUBMITTED",
  MONITORING_REPORT_REJECTED = "MONITORING_REPROT_REJECTED",
  MONITORING_REPORT_APPROVED = "MONITORING_REPORT_APPROVED",
  VERIFICATION_REPORT_SUBMITTED = "VERIFICATION_REPORT_SUBMITTED",
  VERIFICATION_REPORT_REJECTED = "VERIFICATION_REPORT_REJECTED",
  VERIFICATION_REPORT_APPROVED = "VERIFICATION_REPORT_APPROVED",
}